Changed project name in pubspec.yalm. Extracted GradientContainer to seperate file gradient_container.dart.
